Honor AutoTransactionBehavior, move update logic from ClientWrapper to DatabaseWrapper, correct namespace for update logs.
