+++ src/MongoDB.EntityFrameworkCore/Diagnostics/MongoLoggerExtensions.Update.cs
@@ -14,65 +14,22 @@
  */
 
 using System;
-using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.Extensions.Logging;
-using MongoDB.Bson;
 using MongoDB.Driver;
-using MongoDB.EntityFrameworkCore.Query;
 
 namespace MongoDB.EntityFrameworkCore.Diagnostics;
 
-using MqlQueryEventDefinition = EventDefinition<string, CollectionNamespace, string>;
 using BulkWriteEventDefinition = EventDefinition<string, CollectionNamespace, long, long, long>;
 
 /// <summary>
 /// MongoDB-specific logging extensions.
 /// </summary>
-internal static class MongoLoggerExtensions
+internal static partial class MongoLoggerExtensions
 {
-    internal static void ExecutedMqlQuery(
-        this IDiagnosticsLogger<DbLoggerCategory.Database.Command> diagnostics,
-        MongoExecutableQuery mongoExecutableQuery)
-        => ExecutedMqlQuery(diagnostics, mongoExecutableQuery.CollectionNamespace, mongoExecutableQuery.Provider.LoggedStages);
-
-    public static void ExecutedMqlQuery(
-        this IDiagnosticsLogger<DbLoggerCategory.Database.Command> diagnostics,
-        CollectionNamespace collectionNamespace,
-        BsonDocument[]? loggedStages)
-    {
-        var definition = LogExecutedMqlQuery(diagnostics);
-
-        if (diagnostics.ShouldLog(definition))
-        {
-            // Ideally we would always log the query and then log parameter data
-            // when sensitive data is enabled. Unfortunately the LINQ provider
-            // only gives us the query with full values in so we only log MQL
-            // when sensitive logging is enabled.
-            var mql = diagnostics.ShouldLogSensitiveData() ? LoggedStagesToMql(loggedStages) : "?";
-            definition.Log(
-                diagnostics,
-                Environment.NewLine,
-                collectionNamespace,
-                mql);
-        }
-
-        if (diagnostics.NeedsEventData(definition, out var diagnosticSourceEnabled, out var simpleLogEnabled))
-        {
-            var eventData = new MongoQueryEventData(
-                definition,
-                ExecutedMqlQuery,
-                collectionNamespace,
-                LoggedStagesToMql(loggedStages),
-                diagnostics.ShouldLogSensitiveData());
-
-            diagnostics.DispatchEventData(definition, eventData, diagnosticSourceEnabled, simpleLogEnabled);
-        }
-    }
-
     public static void ExecutedBulkWrite(
-        this IDiagnosticsLogger<DbLoggerCategory.Database.Command> diagnostics,
+        this IDiagnosticsLogger<DbLoggerCategory.Update> diagnostics,
         TimeSpan elapsed,
         CollectionNamespace collectionNamespace,
         long documentsInserted,
@@ -108,21 +65,6 @@ internal static class MongoLoggerExtensions
         }
     }
 
-    private static string LoggedStagesToMql(BsonDocument[]? documents)
-        => documents == null
-            ? ""
-            : string.Join(", ", documents.Select(d => d.ToString()));
-
-    private static string ExecutedMqlQuery(EventDefinitionBase definition, EventData payload)
-    {
-        var d = (MqlQueryEventDefinition)definition;
-        var p = (MongoQueryEventData)payload;
-        return d.GenerateMessage(
-            Environment.NewLine,
-            p.CollectionNamespace,
-            p.LogSensitiveData ? p.QueryMql : "?");
-    }
-
     private static string ExecutedBulkWrite(EventDefinitionBase definition, EventData payload)
     {
         var d = (BulkWriteEventDefinition)definition;
@@ -134,29 +76,6 @@ internal static class MongoLoggerExtensions
             p.DocumentsDeleted,
             p.DocumentsModified);
     }
-
-    private static MqlQueryEventDefinition LogExecutedMqlQuery(IDiagnosticsLogger logger)
-    {
-        var definition = ((MongoLoggingDefinitions)logger.Definitions).LogExecutedMqlQuery;
-        if (definition == null)
-        {
-            definition = NonCapturingLazyInitializer.EnsureInitialized(
-                ref ((MongoLoggingDefinitions)logger.Definitions).LogExecutedMqlQuery,
-                logger,
-                static logger => new MqlQueryEventDefinition(
-                    logger.Options,
-                    MongoEventId.ExecutedMqlQuery,
-                    LogLevel.Information,
-                    "MongoEventId.ExecutedMqlQuery",
-                    level => LoggerMessage.Define<string, CollectionNamespace, string>(
-                        level,
-                        MongoEventId.ExecutedMqlQuery,
-                        LogExecutedMqlQueryString)));
-        }
-
-        return (MqlQueryEventDefinition)definition;
-    }
-
 
     private static BulkWriteEventDefinition LogExecutedBulkWrite(IDiagnosticsLogger logger)
     {
@@ -179,8 +98,6 @@ internal static class MongoLoggerExtensions
 
         return (BulkWriteEventDefinition)definition;
     }
-
-    private const string LogExecutedMqlQueryString = "Executed MQL query{newLine}{collectionNamespace}.aggregate([{queryMql}])";
 
     private const string LogExecuteBulkWriteString =
         "Executed Bulk Write ({elapsed} ms) Collection='{collectionNamespace}', Inserted={inserted}, Deleted={deleted}, Modified={modified}";
